Remove erroneous timeout warning
When someone supplies a timeout we check whether that's >= than the default timeout, and if so warn and fall back to the default.  This means that if you set tie timeout to the default you'll get an inappropriate warning.  Change to only warning when it's > than the default.

The only behavior change here is that we're no longer emitting a warning.

Fixes #24163.