Use SameSite=None for canary cookie.
With Chrome 78, cookies will be set with SameSite=Lax by default. This
prevents the canary/rc cookie from being sent when the page is not on
the cdn.ampproject.org domain when requesting v0.js. Use SameSite=None
to set the cookie, working around iOS, which does not treat it correctly
by not setting the SameSite value.

- Include v0.js on the experiments page, so the cookies code can access
the platform service.