shadowDoc: Expose async close method
The existing shadowDoc.close() method is not fully synchronous. Notably,
the 'pass' that occurs on a visibility state change may not complete
before additional tasks are performed, like removing the ShadowRoot host
completely. Cleanup is incomplete and there are some potentially
valuable unload tasks that were not given the chance to run. Expose
shadowDoc.closeAsync() method that waits for the next visibility state
change pass to complete before resolving.