Fix Custom Element instance creation in IE11
When we transpile `super` in Custom Element subclasses, we change it to
`superClass.call(this)` (where `superClass` is `HTMLElementPolyfill`).
That `.call` value is inherited from `Function.prototype`.  But, IE11's
native HTMLElement hierarchy doesn't extend from Function!  And because
`HTMLElementPolyfill` extends from `HTMLElement`, it doesn't have a
`.call`! So we need to manually install it.