shadowDoc: Return promise from close method (#25094)
* shadowDoc: Expose async close method

The existing shadowDoc.close() method is not fully synchronous. Notably,
the 'pass' that occurs on a visibility state change may not complete
before additional tasks are performed, like removing the ShadowRoot host
completely. Cleanup is incomplete and there are some potentially
valuable unload tasks that were not given the chance to run. Expose
shadowDoc.closeAsync() method that waits for the next visibility state
change pass to complete before resolving.

* Add shadowDoc.closeAsync to documentation

* Prefer service/timer promise race method

* Avoid race between resources promise and cleanup

* Revisions based on feedback

* Backwards-compatible shadow close and update unit tests

* Fix comment, remove examples

* Another shot at fixing the comment