Stub out elements that fail to upgrade
I think this is what's causing #25381. If a component, eg `<amp-link-rewritter>`, accesses `this.getAmpDoc()` to get a service in it's constructor, it'll error out.

If that happens, then `this.implementation_` is still `undefined`. That causes a cascade of bugs. So, if the component fails to construct, we're now going to turn it into a stubbed `BaseElement` and pretend it doesn't exist on the page.