Fix transpiled detection in CEv1 polyfill
Closure compiler was eliminating the `opt_ctor.call(instance)` line as useless. This had no real effect on us, because we're transpiling our code.

But, in a world where we shipped native `class` code to browsers, this would have resulted in use wrapping the `HTMLElement` class unnecessarily. Not the full polyfill, but the minimal polyfill.

Fixing this will allow us to eventually use no polyfill on when we ship native classes.