Break infinite loop with loading container
The loading container is created during `updateLayoutBox` (directly after a measure). It uses `mutateElement` to do it, which invalidates the layoutBox. When Resources detects that the layout box is invalid, it remeasures it, and calls `updateLayoutBox`.

Guess what happens?

We call `mutateElement` again. We invalidate the layout box. Resources measures, calls `updateLayoutBox`. On and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on and on...