Delay constructing BaseElement until connectedCallback
This avoids a whole host of problems that can happen innocently inside the implementation's constructor. Eg, trying to get access to the `ampdoc` or a doc-level service.

This also wraps the construction in a try-catch, so that if the instance fails to construct, we will permenently ignore the custom element.