Reduce severity of shadow close timeout log
closeShadowRoot_ was recently updated to return a promise that resolves
on a race between visibility state change completion or 15ms timeout.
The timeout is pretty easy to hit. Remove the sense that a regression
occurred when this feature was introduced due to the new console
warning. Matters have only improved for those taking advantage of the
shadow closure promise. Log at the info level instead of warn.