🐛 inabox: support sending positions to ads with null origins (#25465)
* inabox: support sending positions to ads with null origins

Google AdManager is interested in experimenting with loading AMP Inabox ads with a null origin, for better performance and security.  Before this commit, an inabox ad with a null origin couldn't subscribe to position tracking, because you can't send a postmessage restricted to a null origin.  Reading the code, though, there's no reason to be restricting what origin we send the positions to, because (a) we're just reading the origin from the incoming message and (b) it's completely fine to send positions to any frame that requests them.

Switch to a targetOrigin of '*' to indicate that we have no preference on the receiving origin.

Also add a simple implementation of data-url rendering to the visual tests, so we can verify it works.  Before the change to InaboxMessagingHost the test gives a "Invalid target origin 'null' in a call to 'postMessage'." console error.

* linter

* add comment on why we're not restricting the origin of postMessages

* revert changes to amp-inabox-static.html visual diff test

* fix tests

* update comment