Possible fix for amp-ad beop #25687
It would appear that the confusion would come from the inverted
parameters:

- `requestResize` has (width, height)
- `onResizeDenied` callback has (height, width)

Attaching requestResize to onResizeDenied would invert the height &
width, and given it reads the width from the #c container, mistaking
them results in the resized container to be `InitialHeight x RequestedHeight`.

Took the occasion to only register `onResizeDenied` once instead of
multiple times.