🏗️ Update transform-amp-asserts to leave devAsserts (#26357)
* Update transform-amp-asserts to leave devAsserts

This is necessary for Closure to correctly narrow the type of the value.

If we were to remove the `devAssert(foo)` call, then closure would not be able to tell that `foo` is truthy. This doesn't matter for `assertBoolean`/`assertString`/etc because we're able to replace the call with a typecast. There is no associated "truthy" cast possible for `devAssert`. Same for `dev().assert()`

* Remove unused param