config: Check doc head for custom URLs
Check for a custom URL definition in special <meta> tags. Note that
this does not allow for distinct custom URLs in AmpDocShadow instances.
The shell is allowed to define one set of custom URLs via AMP_CONFIG
(recommended) or by including <meta> tags in the shell <head>. Those
custom URLs then apply to all AMP documents loaded in the shell.