Fix recursion with getting ampdoc
Getting the ampdoc attempts to get the cached value on the node. This is
short-circuited when the node has not yet been attached. If it has been
attached, then it uses `getAmpDoc` on the node. That method will throw
if there's no ampdoc. And because we're just trying to do the lookup to
set the ampdoc, it always failed.