Fix Custom Element instance creation in IE11, redo
This is a continuation of #25245. All of our Custom Elements extend from a common `CustomAmpElement` class, which itself extends from `HTMLElementPolyfill`. In #27470, I removed the `constructor` from `CustomAmpElement` class. But, because it's a subclass and must have a `constructor` that calls `super()`, Closure replaced it with a `constructor() { superClass.apply(this, arguments); }` when transpiling.

But, due to the same faulty inheritance chain described in #25245, `HTMLElementPolyfill` doesn't inherit an `apply` method! So doing `superClass.apply()` throws an error. 😭

This PR installs all important Function methods onto our polyfill, to be resilient to any future changes in the instance creation.