Prevent interaction with blurry-image placeholder (#28090)
The blurry image placeholder uses a transition to slowly fade the blur into the real image. When it's done, the blurry image is "hidden" via `opacity: 0` styles. Unfortunately, users can still interact with this blurry image, and because it overlays the entire image, can't interact with the real image.

Adding `pointer-events: none` seems to be the easiest fix. Other possibilities include adding a `transitionend` event (requires JS, and slow), or converting to a CSS animation (unsure of browser support and smoothness).