🐛 Prevent layoutCallback if resource state changes after scheduling layout (#28849)
* Prevent layoutCallback if resource state changes after scheduling layout

I think there's a race condition where:

1. The element is schedule for layout
2. The page becomes hidden (swipe to next page, or change tabs)
3. The element's scheduled layout finally starts

I think the `state` is being reset to `NOT_LAID_OUT` or `READY_FOR_LAYOUT`, and somehow the `startLayout` method is being called. Either state is acceptable for `startLayout` for some reason. Really, we should only be accepting the resource when it's `LAYOUT_SCHEDULED`, which happens when the (async) layout flow is started. If anything happens between the start of that flow and the time the `startLayout` method is called, we're in a broken flow.

Re: https://github.com/ampproject/amphtml/issues/21116

* Fix tests

* Fix inabox resources