Clone IntersectionObserverEntry for cross-origin postMessages
The `target` property points to a DOM element, which is not structured clonable (a requirement for `postMessage`). We used to `delete entry.target`, to get around this.

But when using native `IntersectionObserver`, the `IntersectionObserverEntry`s has no own properties. Instead, the properties are getters that exist on the prototype. `delete` doesn't work for prototype-inherited properties, so this workaround stopped working. Now, we'll just create a new object with the properties we care about, which is honestly cleaner (you shouldn't mutate objects you don't own).

Fixes https://github.com/ampproject/amphtml/issues/29747