✨ VideoWrapper Preact component with autoplay and mediasession (#29486)
VideoWrapper takes a component to instantiate and control.

This lays between the HTMLMediaElement interface and an outer context component (VideoController or similar), and implements AMP-style autoplay and mediasession. This may not be necessary depending on whether we move up these two features to a future context layer, but advancing this implementation anyhow.