🖍 Cleanup autoplay JSS (#30351)
**Adding `content`**
This slipped me before, and without it `:before/:after` pseudoelements are not displayed.

**Removing `transform`**

Default translate values were used for a workaround where instead of setting `animation-play-state, we'd remove the entire animation`:

https://github.com/ampproject/amphtml/blob/c4a663d0ba13d0488c6fe73c55dc8c971ac6ec0d/src/service/video/autoplay.js#L73-L76

The bug this worked around is no longer [present](https://glitch.com/edit/#!/animation-play-state-transform) in iOS, so we can remove it. If we remove the default transform value, we can also remove the dual selector in `:before, :after {}`.