Add SSR experiments to a4a eid
This adds the currently set SSR experiments to the a4a eid.

A small hiccup with this setup is that the SSR experiments are usually just boolean values (or small enums) that don't have a unique experiment id. I've created a pseudo id by concatenating the proto field id (eg, `25`) with a padded value (eg, `0`) to create the exp id (`2500`).

Re: #30504
Re: cl/335090010