Remove the file extension for script tags in spec_names/messages (`.js`)
This is part 1 of 2.

With the introduction of module/nomodule the file extension for scripts may
be either `.js` or `.mjs` and not always `.js`. This change will allow for the
simplification of handling `also_requires_tag` and `also_requires_tag_warning`.
The latter is used by `amp-ad` and `amp-video`.

The purpose is to allow a single `also_requires_tag_warning` to cover all
variations of script types (regular, module, nomodule, module SXG) with a
single expression (`amp-ad extension script`). The engine will see this and
apply it to any extension script type. That change will come in part 2.

This changes:

`amp-extname extension .js script` to `amp-extname extension script`
`amphtml engine v0.js script` to `amphtml engine script`

PiperOrigin-RevId: 339147096