Alternative tagspecs may meet also_requires_tag_warning requirements
This is part 2 of 2. Part 1 is cl/339147096.

The purpose is to allow a single `also_requires_tag_warning` to cover all
variations of script types (regular, module, nomodule, module SXG) with a
single expression (e.g. `amp-ad extension script`). The engine will see
`extension script` and apply it to any extension script type with the same
extension spec name.

PiperOrigin-RevId: 339898343