Local Server: Listen on all local network addresses
This updates our local server to listen on `0.0.0.0`, which means listening on all available network interfaces. Currently, we're only binding on `127.0.0.1`, which means only the local machine can connect to the server.

By changing to `0.0.0.0`, any device on the local network will be able to connect. This is really benefitial for testing on our phones.

It also prints all of the IPv4 addresses, so that you don't have to go dig that info up.