Fix documentation about amp-list overflow condition
The current documentation seems confusion/misleading - it's saying that
the overflow element **will** display when one of the conditions is met,
meaning that the amp-list will fail to resize if either the bottom of
the amp-list is below the viewport or near the end of the page content.
That seems wrong and opposite of what the actual behavior is: amp-list
should be able to resize and overflow elements should not display if one
of the condition is met.