Allow <amp-img layout=intrinsic> for email spec
The absence of this layout in the spec allowlist seems like an
oversight because the email spec already allows intrinsic layout for
amp-carousel. The SVGs used for sizing intrinsic layout elements are
hardcoded by the AMP runtime, so there's no privacy or security
concerns.

This is discussed with the AMP for Email WG with mail.ru, Yahoo!
Mail and Gmail and the consensus is to allow this.