Allow Service Factories to be adopted by FIE (#32973)
* Allow Service Factories to be adopted by FIE

There are several methods of sharing services with FIE, but the 2 main ones are:

1. FIE installs its own factory
2. FIE copies the parent's factory instance

https://github.com/ampproject/error-reporting/issues/58 is caused because the FIE is using its own factory instance (which means it has its own `Templates` class code, with private names mangled, in each binary that installs the FIE). But each FIE then uses code shared with v0.js to install a amp-mustache instance. This means we have code compiled in v0.js trying to directly interact with a class that's provided by, eg, amp-ad-custom (a separate binary!). This is never good.

What's really needed here is a 3rd way to share services with FIE:

3. FIE copies the parent's **factory**

FIE can then instantiate its own instance of that factory, which will have its own instance data. And because v0.js we'll reuse the class code already in v0.js, we won't have an issue with private mangling. And, we'll make all of these FIE installers smaller!

* Allow disposal of a shared factory's instances

* Fix merge conflict

* Tests

* Update test files

* Update shared instance test