Insert empty ESM export to ensure file is parsed as module
From discussions with the Chrome team, apparently using `<link rel=preload>` and pointing it at a script that's used as `type=module` will cause a double parse of the script. Once as a regular Script, and once as an ESM Module. Adding in an ESM `export {}` ensures that the file cannot be parsed as a Script (because `export` is only valid in ESM Modules), thus saving us from processing this file twice.