Use `default-jdk` instead of `openjdk-7-jre`
Looks like `openjdk-7-jre` [has been removed since Ubuntu 16.04](https://unix.stackexchange.com/a/475657).  Using the `default-jdk` package seems to have worked for me; not sure if there's complete compatibility or not (and my understanding is that it's intended to continually change in the future, so I'm not sure if there's a better approach).