Implementation of multiple tagspecs for a single extension, each with a unique set of version identifiers.
The motivation for this is that Bento AMP components will be using a new version 1.0. With this, we have the opportunity to move forward requirements that these newer version components are actually used in the document, a requirement we have allowed older components to ignore for legacy reasons.

`ExtensionSpec`s are expanded during setup to fill in several fields in the `TagSpec`. One of these is the unique `TagSpec.spec_name` identifier. Multiple extension `TagSpec`s with the same extension name (ie: amp-analytics v1.0 and v0.1) collide on unique `spec_name`.

This change extending the logic for the `ExtensionSpec` -> `TagSpec `generation to include a new field `version_name` which is an optional string that gets added into the `spec_name` and `descriptive_name` for the extension `TagSpec`.

PiperOrigin-RevId: 366894452