♻️  `<amp-ima-video>`: Serialize children (#34229)
`<amp-ima-video>` takes children like:

```
<amp-ima-video>
  <source data-foo="bar" src="src" />
  <track src="src" />
</amp-ima-video>
```

These were previously serialized as HTML on `buildCallback` to be pass to the iframe to (unsafely) write onto `innerHTML`:

```
<amp-ima-video
  data-child-elements='<source data-foo="bar" src="src" /><track src="src" />'
></amp-ima-video>
```

To prepare for a Bento implementation, we serialize into a JSON format that a Preact component can build more simply and safely:

```
[
  ["SOURCE", {"data-foo": "bar", "src": "src"}],
  ["TRACK", {"src": "src"}]
]
```

This change only affects the interface between the `<amp-ima-video>` element and its own proxy `<iframe>`. It does not affect anything author-facing since they will continue to write children `<source>` and `<track>` elements as such.