🚀🏗 ♻️ Remove `assertEnumValue` and minify `isEnumValue` calls (#34465)
We'll often verify that a certain input is a possible enum value. This results in binaries containing something like:

```js
let zd = {
  PRERENDER: "prerender",
  VISIBLE: "visible",
  HIDDEN: "hidden",
  PAUSED: "paused",
  INACTIVE: "inactive",
};
function Ad(a) {
  return J().assertEnumValue(zd, a, "VisibilityState");
}
```

In all of these cases, the enum's keys remain unused. In some cases, the enum object is not required at all.

This change adds a Babel transform to remove them. It also removes `assertEnumValue` in favor of a a composed `assert(isEnumValue())` call, which minifies better.

Notably, the size of `v0.mjs` reduces by `0.12KB` after this change.