🐛 Never show autoplay elements while playing an ad (#34308)
Autoplaying videos display an animated equalizer icon and an overlay unmute button. When we receive `AD_START`, we disable these.

Previously, a race where we receive `AD_START` before we insert the autoplay elements would make them appear on `PLAYING`, before receiving `AD_END`. We now ensure that these elements are not on display when `PLAYING` is received late.