♻️ Updated validator rules
We differentiate versions when adding validator rules with `satisfies`/`excludes` when the rules differ between versions. Since the two versions have `requires_usage` and `deprecated_allow_duplicates` in common (omitted values), their set of validation rules is identical. Therefore we can just do this (effectively the only modification in this file is to insert `version: "1.0"` between `"0.1"` and `"latest"`.

Co-authored-by: Caroline Liu <10456171+caroqliu@users.noreply.github.com>