♻️ Removed `autoplay` for now, `useInOb` will be used in future
To be clear - does `amp-audio:0.1` begin playing immediately once it enters viewport if given autoplay? If so, I think we should consider using the useInOb hook @dmanek is working on to do this feature, perhaps in a separate PR.

Having it `autoplay` immediately is not a good UX, and short of (potentially) implementing the viewport based logic, should be removed from the AMP layer for now.

Co-Authored-By: Alan Orozco <alanorozco@users.noreply.github.com>
Co-Authored-By: Caroline Liu <10456171+caroqliu@users.noreply.github.com>