Pin CircleCI's Chrome version
This is to prevent new stable releases of Chrome from causing test failures. Eg, the Chrome 93 release caused several tests failures that made `main` red until the tests were skipped. This delayed a cherry-pick we're working on, since the new release (on on top of old code) also picked up Chrome 93.

This should be updated regularly, hopefully with a chance to test Chrome releases during the Beta period so we can proactively fix issues.