🏗 Alias `amp-*.js` binaries as `bento-*.js` (#35880)
`amp-*.js` binaries for Bento-compatible extensions provide an `<amp-*>` element that can be used **with** `v0.js`, and a `<bento-*>` element that can be used **without** `v0.js`.

We'd like that Bento elements are used from `bento-*.js` in Bento mode. This unblocks release since we'll be able to include the scripts in Bento mode using the correct URL.

Eventually, this binary will only provide the Bento element, while `amp-*.js` will provide only an AMP element. In the meantime, we alias the multiple-compatibility binary so that `bento-*.js` can be used as soon as possible.