Make amp-form failure message developer-friendlier (#35964)
This can happen whenever the XHR proxy server returns an error and
doesn't include the HTML to render the submit-failure template (e.g.,
AMP CORS failure, unreacheable proxy). Sever-side rendering of template
is an implementation detail of AMP for Email that developers shouldn't
care about so it makes sense to change this to one that would make more
sense for developers.