♻️ `disable-hint-reappear` renamed as `display-hint-once` in Bento `1.0`
The intention is to not have an attribute with a negative name, since enabling or disabling it can be confusing to reason about. On second thought, `repeat-hint` is not a good alternative because it flips the default behaviour when omitted. Let's rename to `display-hint-once` or similar, which has the same default omitted/provided behaviour, but perhaps a clearer name.

Co-Authored-By: Caroline Liu <10456171+caroqliu@users.noreply.github.com>