Update CLS metric to match newest CWV standard (#36147)
* Update CLS metric to match newest CWV standard

The CLS metric used to be a sum of the shifts over the entire lifetime of the page. However, Google has changed this so that it's now defined as the largest sum over a windowed session (see https://web.dev/evolving-cls/).

We're still going to record the old value, but it's now called `cls-1`. Regular `cls` will now be the windowed session value. Additionally, we'll record the types of nodes that contribute to the CLS (as `clstu`, standing for CLS-Type Union), so that we can better understand what's causing it.

* Add comments and descriptive names

* nits

* Fix bug

* Add TICKEVENTS docs

* Cleanup

* Fix tests

* Remove debugging