✨  add new analytics vendor infonline_base to amp project (#36091)
This will add a basic variant of INFOnline analytics to the AMP project.

Why a new analytics vendor?

In Germany, a new law will come into force on December 1st, 2021 (called and abbreviated TTDSG) which massively tightened data protection regulations and forbids any non-technically necessary information processing that takes place without the consent of the user. Since the two other variants infonline_anonymous and infonline each store information on the user's device, this form of information processing is only possible with the consent of the user.

Since INFOnline provides a kind of census measurement for the German advertising market, we are unfortunately forced to register a new analytics vendor for AMP who transmits a minimal amount of information and completely dispenses with the storage of information on the user's device.