Include a random eventid to associate CLS with a page session
In #36147, we switched to the new windowed CLS metric. This new algorithm can send multiple CLS scores, where each new score is larger than the last.

Because of this multiple-sends-which-are-larger-than-the-last behavior, our CLS scores have gone up marginally. The p95 is considering every CLS score
sent by the page, instead of just the most recent. In order to determine which is the most recent, we'll add a random `eventid` to our `sendCsi` message,
which can then be used when aggregating.