🏗 Allow rendering plain DOM using JSX (#36640)
Related to #36459

Provides a replacement for existing rendering methods that's more powerful and convenient:

|                  | JSX | imperative DOM | `htmlFor` | `simple-template` (Stories) |
| ---------------- | --- | -------------- | --------- | --------------------------- |
| like HTML        | ✅  | 🚫             | ✅        | 🚫                          |
| dynamic attrs    | ✅  | ✅             | 🚫        | ✅                          |
| dynamic children | ✅  | 🚫             | 🚫        | sort-of                     |
| event handlers   | ✅  | 🚫             | 🚫        | 🚫                          |
| nesting          | ✅  | 🚫             | 🚫        | 🚫                          |

A minimal implementation of JSX is included as `#core/dom/jsx`.

This library adds ~0.22K of size to any given bundle. By replacing callsites of previous rendering methods, the overall size of a bundle is roughly neutral or smaller.

This change does not change any implementation code. A migration of existing templates and renderers will follow-up.