🏗 🚀 Minify format of locale JSON files (#36662)
JSON locale files used in `amp-story` carry unnecessary information over the wire. Particularly, they may include a `description` field that contains sentences which are only useful for development. We can save a lot of bytes by not stripping this description.

```diff
- {2: {"string": "x", "description": "bla bla"}}
+ {2: {"string": "x"}}
```
This format reduces the compressed size of `amp-story-1.0.mjs` by 1.75%.
We can go further and unwrap the string, which brings the delta to 2.23%:

```diff
- {2: {"string": "x", "description": "bla bla"}}
+ {2: "x"}
```

| format                 | size |       Δ |
| ---------------------- | ------------------: | ------: |
| `{"string": "x", ...}` |             `74261` |     `0` |
| `{"string": "x"}`      |             `72978` | `-1283` |
| `"x"`                  |             `72641` | `-1620` |

This change updates babel-plugin-transform-json-import so that files matching /_locales/ are minified to the unwrapped "x" string format before being inlined.