🏗 Temporarily disable Storybook AMP environment (#36780)
`@ampproject/storybook-addon` has outdated peer dependencies that conflict with the main install. Disable the AMP environment for now so that the Preact environment can run.

We'll update the addon soon: https://github.com/ampproject/storybook-addon-amp/issues/57

Additionally, we were relying on `@storybook/client-api` as a transitive dependency from the addon. Install it so that Storybook can run.