Flush LCP metric once it's received (#37007)
* Flush LCP metric once it's received

LCP isn't a cumulative metric (there's only ever one LCP), so it makes no sense to wait for more metrics to come in. This may be the final reason our CSI metrics don't align with UKM data. LCP pass rate appears low, but it's just because we're not receiving the metric at all even though the LCP already fired.

* Fix test

* Copy comment