♻️🚀 amp-story: Refactor unsupported browser layer for size (#37019)
Refactors Unsupported Browser Layer to minimize bundle size and complexity.

These are compressed size deltas:

### amp-story-1.0.mjs `-0.67kb`

Browsers that run the `module` build support all detected features, so the unsupported browser layer is unnecessary. We strip it out completely.

### amp-story-1.0.js `-0.12kb`

We remove some complexity by encapsulating the layer so that it's cleanly stripped out from the `module` build. As a result, the size of the `nomodule` build also reduces.