Fix first visible timestamp and LCPV (#37144)
In #36273, I switched from `Date.now()` to `performance.now()`, thinking it was a higher resolution timestamp time since the unix epoch. But, `performance.now` is relative to _page load time_ (or `performance.timeOrigin`), not the epoch.

This matters because our "since visible" metrics use the epoch as their base time, meaning the timestamps are roughly 1.6 trillion milliseconds. But `performance.now` is tiny, in the range of a few million for a page that was loaded an hour ago. When we calculated the "since visible", we received values that were still in the trillions when we expect millions.

By adding `performance.now()` and `performance.timeOrigin`, we get a timestamp relative to the epoch. Now when we calculate the "since visible", we get the correct values.

(cherry picked from commit 7645c4edaea449af3d8c8dee8730b0ac98fe6592)