♻️ Fix to avoide multiple request to server
`window` (say global) variable is used to set status of the extension.

Loading of the first most extension initialize parameters and then try to fetch from the server to check AdBlocker.

Subsequent extension only add their functionCallback to `window` parameter.

When first most extension detects blocker, it shows it's own fallback and then call functionCallback of the subsequent extensions.

This save time and resources by sending only one request to the server and based on the responce, all extension on the page shows fallback when AdBlocker is present.