split out generic bento testing from amp tests
current (hacky) workarounds: some tests depend on CSS, which we expect consumers to install. This differs from the AMP components that register styles themselves. To workaround this we have to import and append the bento component styles. This problem might be unique to components that don't use shadowCss. For components that do use shadowDom and have critical styling that needs to be tested, those styles are already written in JSS and declared as shadowCss in the bento component's declaration, which ensures that that the styles are baked into the component.