🐛[bento-facebook] Invert error conditions to make sure both errors are thrown appropriately.  (#37636)
Facebook recently deprecated its Embedded comments feature, and as such we added specific error/warn statements to alert authors to this change. However, `bento-facebook` never actually allowed `embedAs` to equal `comment` and this specific message is unnecessary. 

This PR also fixes the incorrect labeling of the `amp-facebook` warning as `bento-lightbox`.